refactor(electron): split epg event services
